refactor - created searchIndexBuilder page
